requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
google_api_python_client==2.120.0
google_auth_oauthlib==1.2.0
pandas==1.5.2
protobuf==4.25.3
Requests==2.31.0
